bot/requirements.txt
aiogram==3.*
httpx
python-dotenv 